fix recommendations

+++ src/main/java/protocols/dht/messages/InformationGossip.java
@@ -1,20 +1,22 @@
 package protocols.dht.messages;
 
 import pt.unl.fct.di.novasys.network.data.Host;
+
+import java.math.BigInteger;
 import java.util.*;
 
 public class InformationGossip {
     private Map<Integer, Set<Host>> contacts;
     private Set<Host> agView;
-    private Map<Integer, Host> fileTuples;
+    private Map<BigInteger, Host> fileTuples;
 
-    public InformationGossip(Map<Integer, Set<Host>> contacts, Map<Integer, Host> fileTuples, Set<Host> agView) {
+    public InformationGossip(Map<Integer, Set<Host>> contacts, Map<BigInteger, Host> fileTuples, Set<Host> agView) {
         this.contacts = contacts;
         this.agView = agView;
         this.fileTuples = fileTuples;
     }
 
-    public Map<Integer, Host> getFileTuples() {
+    public Map<BigInteger, Host> getFileTuples() {
         return this.fileTuples;
     }
